31/8:update

+++ app/src/main/java/com/tapi/picturesme/core/database/FileDownloadManager.kt
@@ -6,38 +6,34 @@ import android.graphics.Bitmap
 import android.util.Log
 import com.tapi.picturesme.App
 import com.tapi.picturesme.PhotoItem
-import com.tapi.picturesme.core.database.entity.PhotoEntity
-import com.tapi.picturesme.utils.CommonUtils
-import kotlinx.coroutines.Deferred
 import java.io.File
 import java.io.FileOutputStream
 import java.io.IOException
 
 
+fun saveToInternalStorage(bitmapImage: Bitmap, path: String): String? {
 
-     fun saveToInternalStorage(bitmapImage: Bitmap, path: String): String? {
-
-         val cw = ContextWrapper(App.instance.getApplicationContext())
-         // path to /data/data/yourapp/app_data/imageDir
-         val directory: File = cw.getDir("imageDir", Context.MODE_PRIVATE)
-         // Create imageDir
-         val mypath = File(directory, path)
-         var out: FileOutputStream? = null
-         try {
-             out = FileOutputStream(mypath)
-             // Use the compress method on the BitMap object to write image to the OutputStream
-             bitmapImage.compress(Bitmap.CompressFormat.PNG, 100, out)
-         } catch (e: Exception) {
-             e.printStackTrace()
-         } finally {
-             try {
-                 out?.close()
-             } catch (e: IOException) {
-                 e.printStackTrace()
-             }
-         }
-        return directory.getAbsolutePath()
+    val cw = ContextWrapper(App.instance.getApplicationContext())
+    // path to /data/data/yourapp/app_data/imageDir
+    val directory: File = cw.getDir("imageDir", Context.MODE_PRIVATE)
+    // Create imageDir
+    val mypath = File(directory, path)
+    var out: FileOutputStream? = null
+    try {
+        out = FileOutputStream(mypath)
+        // Use the compress method on the BitMap object to write image to the OutputStream
+        bitmapImage.compress(Bitmap.CompressFormat.PNG, 100, out)
+    } catch (e: Exception) {
+        e.printStackTrace()
+    } finally {
+        try {
+            out?.close()
+        } catch (e: IOException) {
+            e.printStackTrace()
+        }
     }
+    return directory.getAbsolutePath()
+}
 
 
     suspend fun isDownloaded(photoItem: PhotoItem): Boolean {
@@ -45,7 +41,9 @@ import java.io.IOException
         for (item in listDB) {
             val link = photoItem.picture.raw
             Log.d("TAG", "isDownloaded11: $link")
+
             val path = link.substring(link.indexOf('-') + 1, link.indexOf('?')) + ".png"
+
             Log.d("TAG", "isDownloaded: $path")
             var pathDb = item.path.substring(item.path.lastIndexOf('/') + 1, item.path.length)
 
@@ -62,9 +60,11 @@ import java.io.IOException
 
     fun deleteFile(path: String): Boolean {
         var file = File(path)
-        file.delete()
-        if (file.delete()) {
-            return true
+        if (file.exists()) {
+            file.delete()
+            if (file.delete()) {
+                return true
+            }
         }
         return false
     }
